Added calibration triggers 8 and 9
Calibration is performed to compute the correct Analog-per-PSI conversion coefficients. Since each circuit reports pressures in linear increments of voltages (i.e. PSI and Analog readings scale linearly), a calibrating simply requires computing the slope for a single data-point. Trigger 8 is responsible for retrieving the user circuit-selection and current circuit-PSI, and then computes the calibration coefficient if the input PSI is nonzero. Trigger 8 will page from the Calibration page to a Calibrated status page displaying the changes made. Trigger 9 will update the GUI with the current circuit-string given the selection as a circuit-index, using a simple circuit-string map.